moving trunks of code back to EthereumNetworkBase.java to keep EthereumNetworkRepository.java  as thin as possible for the convenience of forks
